lib: allow users to use own exception word list which will not change
Through the class `ExceptionWords` users are capable of creating either
a default list of words which should not be changed through the
algorithm or insert their own list by using the 2nd constructor.

As users are forced to pass an object of that class to the algorithm,
users cannot forget to construct such an object.
They can also create multiple lists, in case the user wants to handle
different languages.