lib: build: build library and test artifact separately
This commit introduces 2 new CMakeLists in the subdirs test/ and src/
as it seems to be custom to define a CMakeLists for every new artifact
which seems to be kept in a new directory.

Now, the project root CMakeLists only contains declarations of vars to
be used in the CMakeLists of the subdirs.

Also changed the includes for Catch2 to feel better, as we only use
single_include anyway.