repo: implement automated, extensible git hook infrastructure
Based on Autohook, the current infrastructure is comprised of a single
hooks folder, in which the hook scripts and the autohook script lives.
In this hooks folder, you create more folders, which need to have the
name of the type of hook you want to be run.

You only need you run `autohook_install.sh install` once, and then
create more hook scripts, symlink (or create them directly into the
hook_type folders) and the rest will be handled automatically.

The folder structure should look like this:

repo_root/
├── hooks/
│   ├── autohook_install.sh
│   ├── autohook.sh
│   ├── post-checkout/
│   │   └── 01-delete-pyc-files     # symlink to hooks/scripts/delete-pyc-files.sh
│   ├── pre-commit/
│   │   ├── 01-delete-pyc-files     # symlink to hooks/scripts/delete-pyc-files.sh
│   │   └── 02-run-tests            # symlink to hooks/scripts/run-tests.sh
│   └── scripts/
│       ├── delete-pyc-files.sh
│       └── run-tests.sh
├── other_dirs/
└── other_files